# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ShilpSetu API",
    "description": "One photo. One spoken sentence. A fair-priced, trust-verified listing — no typing, no middleman, no English required. · SIH 2026 · PS ID SIH26090 · Team HACKER LOBBY",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/auth/otp/request": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Request Otp",
        "operationId": "request_otp_api_v1_auth_otp_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpRequested"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/otp/verify": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Verify Otp",
        "operationId": "verify_otp_api_v1_auth_otp_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerify"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "auth"
        ],
        "summary": "Update Me",
        "operationId": "update_me_api_v1_auth_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/capture/uploads": {
      "post": {
        "tags": [
          "capture"
        ],
        "summary": "Create Upload",
        "operationId": "create_upload_api_v1_capture_uploads_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/capture/uploads/{upload_id}": {
      "get": {
        "tags": [
          "capture"
        ],
        "summary": "Upload Status",
        "operationId": "upload_status_api_v1_capture_uploads__upload_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "upload_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Upload Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "capture"
        ],
        "summary": "Put Chunk",
        "description": "Append a chunk at `offset`. A chunk the server already has is ignored (safe to retry after a drop).",
        "operationId": "put_chunk_api_v1_capture_uploads__upload_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "upload_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Upload Id"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/capture/uploads/{upload_id}/complete": {
      "post": {
        "tags": [
          "capture"
        ],
        "summary": "Complete Upload",
        "operationId": "complete_upload_api_v1_capture_uploads__upload_id__complete_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "upload_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Upload Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/capture/uploads/simple": {
      "post": {
        "tags": [
          "capture"
        ],
        "summary": "Simple Upload",
        "description": "Single-request upload for good connections (web kiosk / tests).",
        "operationId": "simple_upload_api_v1_capture_uploads_simple_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "idempotency_key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency Key"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kind"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_simple_upload_api_v1_capture_uploads_simple_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/capture/photo-quality": {
      "post": {
        "tags": [
          "capture"
        ],
        "summary": "Photo Quality",
        "description": "Quick hints for the camera ('too dark', 'move closer'). Advisory only — never blocks capture.",
        "operationId": "photo_quality_api_v1_capture_photo_quality_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_photo_quality_api_v1_capture_photo_quality_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_photo_quality_api_v1_capture_photo_quality_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_photo_quality_api_v1_capture_photo_quality_post"
      },
      "Body_simple_upload_api_v1_capture_uploads_simple_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_simple_upload_api_v1_capture_uploads_simple_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OtpRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "pattern": "^\\+?\\d{10,13}$",
            "title": "Phone",
            "examples": [
              "+919876543210"
            ]
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "hi"
          }
        },
        "type": "object",
        "required": [
          "phone"
        ],
        "title": "OtpRequest"
      },
      "OtpRequested": {
        "properties": {
          "challenge_id": {
            "type": "string",
            "title": "Challenge Id"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "dev_otp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dev Otp",
            "description": "Only returned when SMS_PROVIDER=mock (dev/demo)"
          }
        },
        "type": "object",
        "required": [
          "challenge_id",
          "expires_in"
        ],
        "title": "OtpRequested"
      },
      "OtpVerify": {
        "properties": {
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "code": {
            "type": "string",
            "pattern": "^\\d{6}$",
            "title": "Code"
          },
          "role": {
            "type": "string",
            "enum": [
              "artisan",
              "operator",
              "buyer",
              "admin"
            ],
            "title": "Role",
            "default": "artisan"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          }
        },
        "type": "object",
        "required": [
          "phone",
          "code"
        ],
        "title": "OtpVerify"
      },
...
      "TokenOut": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "role": {
            "type": "string",
            "enum": [
              "artisan",
              "operator",
              "buyer",
              "admin"
            ],
            "title": "Role"
          },
          "has_profile": {
            "type": "boolean",
            "title": "Has Profile"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "user_id",
          "role",
          "has_profile"
        ],
        "title": "TokenOut"
      },
      "UploadCreate": {
        "properties": {
          "idempotency_key": {
            "type": "string",
            "maxLength": 80,
            "minLength": 8,
            "title": "Idempotency Key"
          },
          "kind": {
            "type": "string",
            "enum": [
              "audio",
              "photo"
            ],
            "title": "Kind"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "total_size": {
            "type": "integer",
            "maximum": 26214400.0,
            "exclusiveMinimum": 0.0,
            "title": "Total Size"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          }
        },
        "type": "object",
        "required": [
          "idempotency_key",
          "kind",
          "content_type",
          "total_size"
        ],
        "title": "UploadCreate"
      },
      "UploadStatus": {
        "properties": {
          "upload_id": {
            "type": "string",
            "title": "Upload Id"
          },
          "received": {
            "type": "integer",
            "title": "Received"
          },
          "total_size": {
            "type": "integer",
            "title": "Total Size"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "complete"
            ],
            "title": "Status"
          },
          "storage_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storage Key"
          }
        },
        "type": "object",
        "required": [
          "upload_id",
          "received",
          "total_size",
          "status"
        ],
        "title": "UploadStatus"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}